+++ Sources/StreamChatSwiftUI/ChatChannel/ChannelInfo/FileAttachmentsView.swift
@@ -78,14 +78,13 @@ public struct FileAttachmentsView: View {
                 }
             }
         }
-        .toolbar {
+        .toolbarThemed {
             ToolbarItem(placement: .principal) {
                 Text(L10n.ChatInfo.Files.title)
                     .font(fonts.bodyBold)
                     .foregroundColor(Color(colors.navigationBarTitle))
             }
         }
-        .navigationBarBackground()
         .navigationBarTitleDisplayMode(.inline)
     }
 }