+++ Sources/StreamChatSwiftUI/ChatChannel/ChatChannelView.swift
@@ -138,7 +138,6 @@ public struct ChatChannelView<Factory: ViewFactory>: View, KeyboardReadable {
                     }
                     .opacity(0) // Fixes showing accessibility button shape
                 }
-                .accentColor(colors.tintColor)
                 .overlay(
                     viewModel.reactionsShown ?
                         factory.makeReactionsOverlayView(
@@ -206,6 +205,7 @@ public struct ChatChannelView<Factory: ViewFactory>: View, KeyboardReadable {
         .accessibilityElement(children: .contain)
         .accessibilityIdentifier("ChatChannelView")
         .modifier(factory.makeBouncedMessageActionsModifier(viewModel: viewModel))
+        .accentColor(colors.tintColor)
     }
 
     private var generatingSnapshot: Bool {