+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/Polls/PollAllOptionsView.swift
@@ -43,8 +43,7 @@ struct PollAllOptionsView<Factory: ViewFactory>: View {
                 }
                 .padding()
             }
-            .navigationBarBackground()
-            .toolbar {
+            .toolbarThemed {
                 ToolbarItem(placement: .principal) {
                     Text(L10n.Message.Polls.Toolbar.optionsTitle)
                         .bold()
@@ -57,7 +56,6 @@ struct PollAllOptionsView<Factory: ViewFactory>: View {
                     } label: {
                         Image(systemName: "xmark")
                     }
-                    .accentColor(colors.navigationBarTintColor)
                 }
             }
             .navigationBarTitleDisplayMode(.inline)