+++ StreamChatSwiftUITests/Tests/ChatChannel/ChatChannelView_Tests.swift
@@ -134,9 +134,9 @@ class ChatChannelView_Tests: StreamChatTestCase {
         assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
     }
     
-    func test_chatChannelView_customizedNavigationBar_snapshot() {
+    func test_chatChannelView_themedNavigationBar_snapshot() {
         // Given
-        customizedNavigationBarAppearance()
+        setThemedNavigationBarAppearance()
         let controller = ChatChannelController_Mock.mock(
             channelQuery: .init(cid: .unique),
             channelListQuery: nil,
@@ -157,17 +157,12 @@ class ChatChannelView_Tests: StreamChatTestCase {
         controller.simulateInitial(channel: mockChannel, messages: messages, state: .remoteDataFetched)
 
         // When
-        let view = NavigationView {
-            ScrollView {
-                ChatChannelView(
-                    viewFactory: DefaultViewFactory.shared,
-                    channelController: controller
-                )
-                .frame(width: defaultScreenSize.width, height: defaultScreenSize.height - 64)
-            }
-            .navigationBarTitleDisplayMode(.inline)
-        }
-        .applyDefaultSize()
+        let view = NavigationContainerView {
+            ChatChannelView(
+                viewFactory: DefaultViewFactory.shared,
+                channelController: controller
+            )
+        }.applyDefaultSize()
 
         // Then
         assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))